Adding author attribution for YouTube
This experimental feature will attempt to use the blog user's account
when it's available. This has the advantage of keeping the blog user's
data consistent and will benefit blogs that have custom callouts, etc.
for the post author.

This commit also adds a rel="nofollow" to the comment source
attribution.